codebase_scanner/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database and ORM
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and analysis
anthropic==0.7.7

# Background tasks
celery==5.3.4
redis==5.0.1

# Security scanning tools
semgrep==1.52.0
bandit==1.7.5
safety==2.4.0

# Git operations
GitPython==3.1.40

# Data processing
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.2
pydantic-settings==2.1.0

# File processing
python-magic==0.4.27
zipfile38==0.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.39.2

# Utils
requests==2.31.0
click==8.1.7
typing-extensions==4.8.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1